Ensure feedback forms are not duplicated and are easily joined in feedback manager.
The mg_feedback_id field is used to join the survey to the first feedback form. The "unique" id is currently using a timestamp and an added number to it. I think the original intention was to append a random number to it. But previously it left open the chance to join two wrong entires. Now it will use the formstack unique ID returned from the first form to populate it in the second form.

The feedback manager ETL will take care of the mg_feedback_id field on the first form when it executes the formstack webhook.

Currently you can double click the feedback form submission buttons and submit multiple submissions to formstack. In the case of the survey I was able to submit 5 submissions by clicking quickly.

Added some validation checks that should stop a robot submitting forms.