DP-34727: Topic page flags and features (#2654)
* config changes

* DP-34727: Add conditional feedback form rendering.

Introduced `TopicPageFormAlter` to control access to topic fields on the form. Updated the feedback template to conditionally include the feedback form based on `field_hide_feedback_component` value.

* DP-34727: Fix field visibility conditions for topic_page form display

Update the configuration of the topic_page form display to ensure that the depended field's state and condition are correctly set. This change ensures that the form behaves as expected based on the organizational restrictions.

* DP-34727: Refactor topic field access control logic

Streamline the access control logic for topic fields by looping through a predefined set of fields. This approach reduces redundancy and ensures consistent access checks based on user permissions.

* DP-34727: Add Change log.

* DP-34727: Add more documentation.

* DP-34727: Fix yml.

* DP-34727: Conditionally display related organizations and topics section

Wrapped the related organizations and topics section in a conditional statement. This ensures the section is only rendered if there are related organizations or topics present.

* DP-34727: Show the field_organizations for not admins, but keep it disabled.

* DP-34727: Refactor topic fields access control

Separate access control logic into dedicated methods for better readability and maintainability. Introduce `adminFieldsAccess` and `organizationsFieldAccess` to handle specific field access rules based on user permissions.

* DP-34727: Update feedback component description in topic_page field

Clarified that the feedback form is always hidden if no organization is selected. This change aims to prevent confusion for content administrators.

* DP-34727: Variable name.

* DP-34727: Refactor TopicPageFormAlter for enhanced field access control

Update the `topicFieldsControlAccess` method to accept `FormStateInterface` and incorporate extended access control logic for `field_organizations` based on `field_restrict_orgs_field`. This refactor ensures that the access control depends on specific node-level permissions and field values.

* DP-34727: Disable conditional_fields form validation when organization field is restricted

Remove conditional field validations when the organization field is restricted by admin settings. This ensures unnecessary validations are not applied, streamlining the form submission process.

* DP-34727: Place the test to right directory.

* DP-34727: Remove unused "use" statement.

* DP-34727: Prevent link management if restricted globally

Add a condition to bypass link management modifications in `TopicPageFormAlter` if globally restricted for the Topic Page. This ensures that no changes are made to link management when the restriction is active.

* DP-34727: Add tests for feedback form visibility on topic pages

Introduce a method to create topic pages with and without feedback forms and add corresponding tests to verify the visibility of the feedback form based on the provided condition. This enhances the coverage and robustness of the Topic Page functionality by ensuring the feedback form behaves as expected.

* DP-34727: Test admin controlled fields visibility on Topic Page.

* DP-34727: Add more checks.

* DP-34727: Tests access to organizations field on topic page edit form.

* DP-34727: Code styles.

* DP-34727: Add test for Topic Page admin field behavior

Introduces `TopicPageAdminFieldsTest` to verify the behavior of the organization field on the topic page. Ensures the field is disabled when the related control option is selected by an admin.

* DP-34727: Check non-admin/admin can save the form.

* DP-34727: Code validator fixes.

---------

Co-authored-by: Joe Galluccio <joseph.galluccio@state.ma.us>
Co-authored-by: Dima Storozhuk <dmutro.storozhuk@gmail.com>