DP-38375 image optimization for featured item mosaic (#2946)
* DP-38375: Refactor image rendering in featured item templates.

Replaced custom image rendering logic with `drupal_image()` for consistent image processing and responsive support. Simplified template logic and removed unnecessary image styles in the theme file to streamline functionality.

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>

* DP-38375: Add changelog entry for responsive images update

Added a changelog file detailing the implementation of responsive images in specific Twig templates. This ensures proper documentation for the changes made under issue DP-38375.

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>

* DP-38375: Update redirect to PL.

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>

* DP-38375: Fix spacing in 'responsive' parameter in featured item templates

Adjusted the spacing around the 'responsive' parameter in multiple Twig templates to ensure consistency and adhere to coding standards. This change does not impact functionality but improves code readability and formatting.

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>

* DP-38375: Update mayflower-artifacts reference in composer.lock

Switched to the feature branch `DP-38375-Optimize-images-for-image-paragraph-on-info-detail-pages`. This ensures the latest changes for optimizing images are included in the project. Updated metadata such as commit reference and timestamp accordingly.

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>

* DP-38375: Refactor image rendering to use responsive_image fields.

Updated Twig templates and configurations to replace drupal_image function with responsive_image fields for better compatibility and maintainability. Ensured usage of the 'full_width' responsive image style across all instances.

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>

* DP-38375: Update mayflower-artifacts dependency to latest version

Updated the composer.lock to point to the latest reference of mayflower-artifacts. This ensures the project uses the most recent updates and fixes from the dependency repository.

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>

* Mayflower

* Composer

* Composer

* Composer

---------

Signed-off-by: Dmytro Storozhuk <dima@itech4web.com>
Co-authored-by: arthurbaghdas <arthur.baghdasar@gmail.com>